Modified LocationSurrogate.java to use complete address list
BUG: Prevous version used a random generator function (Utility.GetRandomChars) that could only generate items that start with 2 letters.
FIX: I added a new version of the function(GetRandomChars2) that can also generate numbers, spaces, and punctuation marks that can occur in addresses, so that the address anonymizer can select from any address in the address list (address_g3.csv).

MODIFICATION: The previous version was also limited to only use the top 10 most frequent addresses for each set of addresses that begin with the same 2 characters. I modified this to remove this restriction.
